feat: fix issue login

+++ src/lib/axios.ts
@@ -24,7 +24,12 @@ apiClient.interceptors.request.use(
 apiClient.interceptors.response.use(
   (response) => response,
   (error) => {
-    if (error.response?.status === 401) {
+    console.log("axios", error.response);
+
+    if (
+      error.response?.status === 401 &&
+      !error.config.url?.includes("/auth/login")
+    ) {
       localStorage.removeItem("userData");
       store.dispatch(removeToken());
       window.location.href = "/login";
